Extract lots of MutableShow-related classes to their own files.
