+++ src/commonMain/kotlin/baaahs/show/mutable/EditHandler.kt
@@ -1,0 +1,7 @@
+package baaahs.show.mutable
+
+interface EditHandler<T, TState> {
+    fun onEdit(mutableDocument: MutableDocument<T>, pushToUndoStack: Boolean = true)
+    fun onEdit(document: T, pushToUndoStack: Boolean = true)
+    fun onEdit(document: T, documentState: TState, pushToUndoStack: Boolean = true, syncToServer: Boolean = true)
+}
