+++ src/commonMain/kotlin/baaahs/show/mutable/MutableShow.kt
@@ -2,10 +2,13 @@ package baaahs.show.mutable
 
 import baaahs.SparkleMotion
 import baaahs.app.ui.dialog.DialogPanel
-import baaahs.app.ui.editor.*
-import baaahs.control.*
+import baaahs.app.ui.editor.EditableManager
+import baaahs.app.ui.editor.PatchEditorPanel
+import baaahs.control.MutableButtonControl
+import baaahs.control.MutableButtonGroupControl
+import baaahs.control.OpenButtonControl
+import baaahs.control.OpenButtonGroupControl
 import baaahs.getBang
-import baaahs.getValue
 import baaahs.gl.Toolchain
 import baaahs.gl.openShader
 import baaahs.gl.patch.ContentType
@@ -19,140 +22,6 @@ import baaahs.show.live.OpenPatchHolder
 import baaahs.show.live.OpenShow
 import baaahs.show.live.ShaderInstanceResolver
 import baaahs.util.CacheBuilder
-import baaahs.util.UniqueIds
-
-interface EditHandler<T, TState> {
-    fun onEdit(mutableDocument: MutableDocument<T>, pushToUndoStack: Boolean = true)
-    fun onEdit(document: T, pushToUndoStack: Boolean = true)
-    fun onEdit(document: T, documentState: TState, pushToUndoStack: Boolean = true, syncToServer: Boolean = true)
-}
-
-abstract class MutablePatchHolder(
-    private val basePatchHolder: PatchHolder
-) : MutableEditable<Show> {
-    protected abstract val mutableShow: MutableShow
-
-    override var title = basePatchHolder.title
-
-    override fun getEditorPanels(editableManager: EditableManager<*>): List<DialogPanel> {
-        return listOf(
-            GenericPropertiesEditorPanel(editableManager, getPropertiesComponents()),
-            PatchHolderEditorPanel(editableManager, this)
-        )
-    }
-
-    open fun getPropertiesComponents(): List<PropsEditor> {
-        return listOf(TitlePropsEditor(this))
-    }
-
-    val patches by lazy {
-        basePatchHolder.patches.map { MutablePatch(it, mutableShow) }.toMutableList()
-    }
-    val eventBindings = basePatchHolder.eventBindings.toMutableList()
-
-    internal val controlLayout by lazy {
-        basePatchHolder.controlLayout
-            .map { (panelId, controlIds) ->
-                mutableShow.findPanel(panelId) to
-                        controlIds.map { mutableShow.findControl(it) }.toMutableList()
-            }.toMap(mutableMapOf())
-    }
-
-    fun addPatch(block: MutablePatch.() -> Unit): MutablePatchHolder {
-        val mutablePatch = MutablePatch(
-            emptyList(),
-            Surfaces.AllSurfaces
-        )
-        mutablePatch.block()
-        addPatch(mutablePatch)
-        return this
-    }
-
-    fun addPatch(mutablePatch: MutablePatch): MutablePatchHolder {
-        val existingPatch = patches.find { it.surfaces == mutablePatch.surfaces }
-        if (existingPatch != null) {
-            existingPatch.mutableShaderInstances.addAll(mutablePatch.mutableShaderInstances)
-            if (SparkleMotion.EXTRA_ASSERTIONS) existingPatch.mutableShaderInstances.assertNoDuplicates()
-        } else {
-            patches.add(mutablePatch)
-        }
-        return this
-    }
-
-    fun editPatch(index: Int, block: MutablePatch.() -> Unit): MutablePatchHolder {
-        patches[index].block()
-        return this
-    }
-
-    fun addControl(panel: MutablePanel, control: MutableControl) {
-        controlLayout.getOrPut(panel) { arrayListOf() }.add(control)
-    }
-
-    fun addButton(panel: MutablePanel, title: String, block: MutableButtonControl.() -> Unit): MutableButtonControl {
-        val control = MutableButtonControl(ButtonControl(title), mutableShow)
-        control.block()
-        addControl(panel, control)
-        return control
-    }
-
-    fun addButtonGroup(
-        panel: MutablePanel,
-        title: String,
-        direction: ButtonGroupControl.Direction = ButtonGroupControl.Direction.Horizontal,
-        block: MutableButtonGroupControl.() -> Unit
-    ): MutableButtonGroupControl {
-        val control = MutableButtonGroupControl(title, direction, mutableListOf(), mutableShow)
-        control.block()
-        addControl(panel, control)
-        return control
-    }
-
-    fun addVacuity(
-        panel: MutablePanel
-    ): MutableVacuityControl {
-        val control = MutableVacuityControl("Vacuity")
-        addControl(panel, control)
-        return control
-    }
-
-    fun removeControl(panel: Panel, index: Int): MutableControl {
-        return controlLayout.getOrPut(mutableShow.findPanel(panel)) { arrayListOf() }.removeAt(index)
-    }
-
-    fun findControlDataSources(): Set<DataSource> {
-        return controlLayout.values.flatMap {
-            it.filterIsInstance<MutableDataSourcePort>().map { it.dataSource }
-        }.toSet()
-    }
-
-    fun editControlLayout(panel: Panel): MutableList<MutableControl> {
-        return controlLayout.getOrPut(mutableShow.findPanel(panel)) { mutableListOf() }
-    }
-
-    internal fun buildControlLayout(showBuilder: ShowBuilder): Map<String, List<String>> {
-        return controlLayout.map { (panel, controls) ->
-            showBuilder.idFor(panel.build()) to
-                    controls.map { mutableControl ->
-                        mutableControl.buildAndStashId(showBuilder)
-                    }
-        }.toMap()
-    }
-
-    open fun accept(visitor: MutableShowVisitor, log: VisitationLog = VisitationLog()) {
-        if (log.patchHolders.add(this)) visitor.visit(this)
-        patches.forEach { it.accept(visitor, log) }
-        controlLayout.forEach { (_, controls) ->
-            controls.forEach { it.accept(visitor, log) }
-        }
-    }
-}
-
-interface MutableDocument<T> : MutableEditable<T> {
-    fun build(): T
-
-    fun isChanged(originalDocument: T): Boolean =
-        originalDocument != build()
-}
 
 class MutableShow(
     baseShow: Show
@@ -444,69 +313,4 @@ data class MutableShaderInstance(
             } == true
         }
     }
-}
-
-
-class ShowBuilder {
-    private val panelIds = UniqueIds<Panel>()
-    private val controlIds = UniqueIds<Control>()
-    private val dataSourceIds = UniqueIds<DataSource>()
-    private val shaderIds = UniqueIds<Shader>()
-    private val shaderInstanceIds = UniqueIds<ShaderInstance>()
-
-    fun idFor(panel: Panel): String {
-        return panelIds.idFor(panel) { panel.suggestId() }
-    }
-
-    fun idFor(control: Control): String {
-        return controlIds.idFor(control) { control.suggestId() }
-    }
-
-    fun idFor(dataSource: DataSource): String {
-        return dataSourceIds.idFor(dataSource) { dataSource.suggestId() }
-    }
-
-    fun idFor(shader: Shader): String {
-        return shaderIds.idFor(shader) { shader.suggestId() }
-    }
-
-    fun idFor(shaderInstance: ShaderInstance): String {
-        return shaderInstanceIds.idFor(shaderInstance) { "${shaderInstance.shaderId}-inst" }
-    }
-
-    fun getControls(): Map<String, Control> = controlIds.all()
-    fun getDataSources(): Map<String, DataSource> = dataSourceIds.all()
-    fun getShaders(): Map<String, Shader> = shaderIds.all()
-    fun getShaderInstances(): Map<String, ShaderInstance> = shaderInstanceIds.all()
-
-    // Make sure we include data source dependencies, otherwise their feeds aren't opened.
-    // This is pretty janky, find a better way.
-    fun includeDependencyDataSources() {
-        getDataSources().forEach { (_, dataSource) ->
-            dataSource.dependencies.forEach { (_, dependency) ->
-                idFor(dependency)
-            }
-        }
-    }
-
-    companion object {
-        fun forImplicitControls(
-            existingControls: Map<String, Control>,
-            existingDataSources: Map<String, DataSource>
-        ): ShowBuilder = ShowBuilder().apply {
-            controlIds.putAll(existingControls)
-            dataSourceIds.putAll(existingDataSources)
-        }
-    }
-}
-
-class VisitationLog {
-    val patchHolders = mutableSetOf<MutablePatchHolder>()
-    val patches = mutableSetOf<MutablePatch>()
-    val surfaces = mutableSetOf<Surfaces>()
-    val shaderInstances = mutableSetOf<MutableShaderInstance>()
-    val shaders = mutableSetOf<MutableShader>()
-    val shaderChannels = mutableSetOf<MutableShaderChannel>()
-    val controls = mutableSetOf<MutableControl>()
-    val dataSources = mutableSetOf<MutableDataSourcePort>()
 }
